refactor: change db file configuration
