+++ routes/list.tsx
@@ -1,7 +1,8 @@
 import { type Handlers, type PageProps } from "$fresh/server.ts";
 import { desc, sql } from "drizzle-orm";
 import { db } from "~/core/db.ts";
-import { movieTable, type PickMovie } from "~/db/schema/movie.ts";
+import { movieTable } from "~/db/schema.ts";
+import type { PickMovie } from "~/db/model.ts";
 import { Heading } from "~/components/atoms/Heading.tsx";
 import { Layout } from "~/components/organisms/Layout.tsx";
 import { MovieCardList } from "~/components/organisms/MovieCardList.tsx";
