+++ routes/search.tsx
@@ -2,7 +2,8 @@ import { type Handlers, type PageProps } from "$fresh/server.ts";
 import { desc, like, sql } from "drizzle-orm";
 import { getUrlParams } from "~/core/api.ts";
 import { db } from "~/core/db.ts";
-import { movieTable, type PickMovie } from "~/db/schema/movie.ts";
+import { movieTable } from "~/db/schema.ts";
+import type { PickMovie } from "~/db/model.ts";
 import { Heading } from "~/components/atoms/Heading.tsx";
 import { Layout } from "~/components/organisms/Layout.tsx";
 import { MovieCardList } from "~/components/organisms/MovieCardList.tsx";
