+++ routes/theater/add.ts
@@ -3,7 +3,8 @@ import { getCookies } from "$std/http/cookie.ts";
 import { getUrlParams, redirectResponse } from "~/core/api.ts";
 import { db } from "~/core/db.ts";
 import { isInvalidAccount } from "~/core/util.ts";
-import { type NewTheater, theaterTable } from "~/db/schema/theater.ts";
+import { theaterTable } from "~/db/schema.ts";
+import type { NewTheater } from "~/db/model.ts";
 
 export const handler: Handlers = {
   async POST(req) {
